Poprawki modeli - wy...walenie siedliska bugów zwanego słownikami.
