+++ SerwisAudiometryczny/SerwisAudiometryczny/Interface/IBaseModel.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SerwisAudiometryczny.Interface
+{
+    interface IMyBase
+    {
+        int ID
+        {
+            get;
+            set;
+        }
+    }
+}
