+++ SerwisAudiometryczny/SerwisAudiometryczny/Models/InstrumentModel.cs
@@ -1,11 +1,12 @@
-﻿using System;
+﻿using SerwisAudiometryczny.Interface;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
 namespace SerwisAudiometryczny.Models
 {
-    public class InstrumentModel
+    public class InstrumentModel : BaseModel, IMyBase
     {
         /// <summary>
         /// Nazwa instrumentu.
